fix: work around Sending request failed
Turns out it was an error with how the worker was being built.

Failed to execute 'postMessage' on 'Worker': has(D){return typeof h(c,D)!="undefined"} could not be cloned.

This along with some debugging through to vscode-jsonrpc put me on
the right track to look closer at the initializationParams. Two quick
Go to Definitions later and I found the smoking gun.
WorkspaceConfiguration.has.