fix: resolve the real path to links in the parser
Like with reference finding, we may have links that point to
files that are on disk, but behind a symlink. Previously we'd
end up with two separate ScssDocument instances. One with the
real path as key, and one with the node_modules/symlinked path.

With this change developers in monorepos should see their changes
in a symlinked dependency be reflected without a reload. Hopefully
memory hasn't been an issue, but usage should be lower in any case.