refactor: reorganize the repo as an npm workspace
Since we'll be publishing the language server as a standalone
thing it's about time we rearranged some things. The goal is
to better show the language server for visitors to the repo,
while at the same time keeping the development workflow for the
server and vscode client (colocated for debugging, e2e testing).