refactor: rewrite the internals of the language server
The internals of the language server are now more uniform. All parts
are working on the same parsed data structure instead of being a mix
of three different parsers.

This refactor includes a private fork of vscode-css-languageservice.
It holds the parser and link resolution, and some language features
we either extend, modify, or forward. It's kept as a separate package
to hopefully make it easier to keep up with upstream, and to send
patches upstream as well.

A goal with this refactor is to make it easier to maintain the codebase
both for myself and for new contributors.