feat: add two settings to tweak completions in other editors (#224)
Testing in Helix, which uses a different grammar for SCSS than Code,
revealed a few annoying nits with compleitons.

This version adds two new settings that can be useful for maintainers
of language clients and editor configurations:

- `somesass.completion.afterModule`
- `somesass.completion.beforeVariable`

`afterModule` lets you control whether or not to insert a `.` when
accepting a suggestion from modules. If you see for instance
`module..$variable` then try setting `afterModule` to an empty string.

`beforeVariable` is for non-module variables. If you see double dollars
(`$$variable`) then set `beforeVariable` to the empty string.